Correccion de controladores, modelos y validaciones

+++ Backend/src/controllers/userController.js
@@ -28,16 +28,14 @@ export const createUser = async (req, res) => {
       return handleError(res, null, session, 409, 'El DNI ya está registrado');
     }
 
-    // Si el correo es null o está vacío, asigna el DNI al campo email
-    let userEmail = email && email.trim() !== '' ? email : dni;
+    let userEmail = email && email.trim() !== '' ? email : null;
 
-    // Verifica si el correo ya está registrado, excepto si el email es una cadena vacía
     if (userEmail) {
       const emailExists = await User.exists({ email: userEmail });
       if (emailExists) {
         return handleError(res, null, session, 409, 'El correo electrónico ya está registrado');
       }
-    } 
+    }
 
     const roleDoc = await Role.findById(role);
 
@@ -56,10 +54,11 @@ export const createUser = async (req, res) => {
       passwordHash = await encryptPassword(password);
     }
 
-    const newUser = new User({ name, lastName, dni, email, role, isActive });
+    const newUser = new User({ name, lastName, dni, email: userEmail || '', role, isActive });
+
     await newUser.save({ session });
 
-    const newCredential = new Credential({ email: userEmail, password: passwordHash, user: newUser._id });
+    const newCredential = new Credential({ email: userEmail || dni, password: passwordHash, user: newUser._id });
     await newCredential.save({ session });
 
     // await saveAuditEntry({
@@ -100,18 +99,34 @@ export const createUser = async (req, res) => {
 // Obtener todos los Usuarios
 export const getUsers = async (req, res) => {
   try {
-
     const users = await User.find()
       .select("_id name lastName dni email isActive role")
       .populate({
         path: 'role',
         select: '_id name'
-      }).exec();
+      })
+      .exec();
 
     if (!users.length) {
       return res.status(404).json({ error: 'No existen usuarios' });
     }
-    res.status(200).json({ data: users, message: "Usuarios extraídos con éxito" });
+
+    const userIds = users.map(user => user._id);
+    const credentials = await Credential.find({ user: { $in: userIds } })
+      .select('user isActive')
+      .exec();
+
+    const credentialsMap = credentials.reduce((acc, credential) => {
+      acc[credential.user] = credential.isActive;
+      return acc;
+    }, {});
+
+    const usersWithCredentialStatus = users.map(user => ({
+      ...user.toObject(),
+      credentialStatus: credentialsMap[user._id] || null
+    }));
+
+    res.status(200).json({ data: usersWithCredentialStatus, message: "Usuarios extraídos con éxito" });
   } catch (error) {
     handleError(res, error);
   }
@@ -175,10 +190,8 @@ export const updateUser = async (req, res) => {
       return handleError(res, null, session, 404, 'Credencial no encontrada');
     }
 
-    // Actualiza el usuario
     const updatedUser = await User.findByIdAndUpdate(id, { name, lastName, dni, email, role, isActive }, { new: true, session }).exec();
 
-    // Solo actualiza la credencial si el email no es una cadena vacía
     if (email && email.trim() !== '') {
       await Credential.findByIdAndUpdate(credential._id, { email }, { new: true, session }).exec();
     }
@@ -220,12 +233,11 @@ export const updateUser = async (req, res) => {
   }
 };
 
-
-
 // Eliminar Usuario
 export const deleteUser = async (req, res) => {
   let session;
   try {
+
     session = await mongoose.startSession();
     session.startTransaction();
 
@@ -236,12 +248,20 @@ export const deleteUser = async (req, res) => {
       return handleError(res, null, session, 404, "Usuario no encontrado");
     }
 
-    await User.findByIdAndDelete(id, { session }).exec();
+    const credential = await Credential.findOne({ user: id }).exec();
+
+    if (!credential) {
+      await session.abortTransaction();
+      return handleError(res, null, session, 404, 'Credencial no encontrada');
+    }
 
     if (await Client.exists({ user: id })) {
       await session.abortTransaction();
       return res.status(409).json({ error: "Existe un cliente asociados a este usuario" });
     }
+
+    await User.findByIdAndDelete(id, { session }).exec();
+    await Credential.deleteOne({user: id}, { session }).exec();
 
     // await saveAuditEntry({
     //   eventType: 'DELETE',
@@ -261,6 +281,7 @@ export const deleteUser = async (req, res) => {
         console.error('Error al abortar la transacción:', abortError);
       }
     }
+    console.log(error)
     handleError(res, error, session, 500, "Error al eliminar el usuario");
   } finally {
     if (session) {
